pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["abdallah alabed <abd.alabed97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.12.4"
pytz = "^2021.3"
psycopg2-binary = "^2.9.2"
djangorestframework-simplejwt = "^5.0.0"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
